Repository and remote renames
RepositoryForm -> ContainerRepositoryForm

DeleteRemoteModal
