+++ src/actions/ansible-remote-delete.tsx
@@ -1,7 +1,7 @@
 import { msg, t } from '@lingui/macro';
 import React from 'react';
 import { AnsibleRemoteAPI } from 'src/api';
-import { DeleteAnsibleRemoteModal } from 'src/components';
+import { DeleteRemoteModal } from 'src/components';
 import {
   handleHttpError,
   parsePulpIDFromURL,
@@ -14,7 +14,7 @@ export const ansibleRemoteDeleteAction = Action({
   title: msg`Delete`,
   modal: ({ addAlert, listQuery, setState, state }) =>
     state.deleteModalOpen ? (
-      <DeleteAnsibleRemoteModal
+      <DeleteRemoteModal
         closeAction={() => setState({ deleteModalOpen: null })}
         deleteAction={() =>
           deleteRemote(state.deleteModalOpen, { addAlert, setState, listQuery })
