+++ src/api/ansible-remote.ts
@@ -1,5 +1,38 @@
 import { PulpAPI } from './pulp';
-import { type AnsibleRemoteType } from './response-types/ansible-remote';
+
+export class AnsibleRemoteType {
+  auth_url: string;
+  ca_cert: string;
+  client_cert: string;
+  download_concurrency: number;
+  name: string;
+  proxy_url: string;
+  pulp_href?: string;
+  rate_limit: number;
+  requirements_file: string;
+  tls_validation: boolean;
+  url: string;
+  signed_only: boolean;
+  sync_dependencies?: boolean;
+
+  // connect_timeout
+  // headers
+  // max_retries
+  // policy
+  // pulp_created
+  // pulp_labels
+  // pulp_last_updated
+  // sock_connect_timeout
+  // sock_read_timeout
+  // total_timeout
+
+  hidden_fields: {
+    is_set: boolean;
+    name: string;
+  }[];
+
+  my_permissions?: string[];
+}
 
 // simplified version of smartUpdate from execution-environment-registry
 function smartUpdate(
