+++ src/components/collection-dependencies-list.tsx
@@ -4,7 +4,6 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { type CollectionVersion, type CollectionVersionSearch } from 'src/api';
 import { EmptyStateNoData, HelpButton } from 'src/components';
-import 'src/containers/collection-detail/collection-dependencies.scss';
 
 interface IProps {
   collection: CollectionVersionSearch;
