+++ src/components/collection-usedby-dependencies-list.tsx
@@ -10,7 +10,6 @@ import {
   PulpListToolbar,
   PulpPagination,
 } from 'src/components';
-import 'src/containers/collection-detail/collection-dependencies.scss';
 import { Paths, formatPath } from 'src/paths';
 import { ParamHelper, filterIsSet } from 'src/utilities';
