+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export { AccessTab } from './access-tab';
 export { AlertList, type AlertType, closeAlert } from './alert-list';
-export { AnsibleRepositoryForm } from './ansible-repository-form';
 export { AppliedFilters } from './applied-filters';
 export { ApprovalRow } from './approval-row';
 export { ApproveModal } from './approve-modal';
@@ -19,17 +18,18 @@ export { CollectionListItem } from './collection-list-item';
 export { CollectionUsedbyDependenciesList } from './collection-usedby-dependencies-list';
 export { CompoundFilter, type FilterOption } from './compound-filter';
 export { ConfirmModal } from './confirm-modal';
+export { ContainerRepositoryForm } from './container-repository-form';
 export { CopyCollectionToRepositoryModal } from './copy-collection-to-repository-modal';
 export { CopyURL } from './copy-url';
 export { DarkmodeSwitcher } from './darkmode-switcher';
 export { DataForm } from './data-form';
 export { DateComponent } from './date-component';
-export { DeleteAnsibleRemoteModal } from './delete-ansible-remote-modal';
 export { DeleteAnsibleRepositoryModal } from './delete-ansible-repository-modal';
 export { DeleteCollectionModal } from './delete-collection-modal';
 export { DeleteExecutionEnvironmentModal } from './delete-execution-environment-modal';
 export { DeleteGroupModal } from './delete-group-modal';
 export { DeleteModal } from './delete-modal';
+export { DeleteRemoteModal } from './delete-remote-modal';
 export { DeleteUserModal } from './delete-user-modal';
 export { DeprecatedTag } from './deprecated-tag';
 export { DetailList } from './detail-list';
@@ -78,6 +78,7 @@ export { NamespaceListItem } from './namespace-list-item';
 export { NamespaceModal } from './namespace-modal';
 export { NotFound } from './not-found';
 export { CollectionNumericLabel } from './numeric-label';
+export { OrphanCleanupTaskModal } from './orphan-cleanup-task-modal';
 export { Page } from './page';
 export { PageWithTabs } from './page-with-tabs';
 export { PartnerHeader } from './partner-header';
@@ -109,7 +110,9 @@ export { PulpDownloadButton } from './pulp-download-button';
 export { PulpLabels } from './pulp-labels';
 export { PulpListToolbar } from './pulp-list-toolbar';
 export { PulpPagination } from './pulp-pagination';
+export { PurgeTaskModal } from './purge-task-modal';
 export { RemoteForm } from './remote-form';
+export { RepairTaskModal } from './repair-task-modal';
 export { RepositoryBadge } from './repository-badge';
 export { RepositoryForm } from './repository-form';
 export { ResourcesForm } from './resources-form';
