+++ src/components/pulp-code-block.tsx
@@ -6,8 +6,7 @@ import {
   ExpandableSection,
 } from '@patternfly/react-core';
 import React from 'react';
-import { PulpCopyButton } from 'src/components/pulp-copy-button';
-import { PulpDownloadButton } from 'src/components/pulp-download-button';
+import { PulpCopyButton, PulpDownloadButton } from 'src/components';
 
 export const PulpCodeBlock = ({
   code,
