+++ src/components/purge-task-modal.tsx
@@ -8,7 +8,7 @@ import {
   Modal,
 } from '@patternfly/react-core';
 import React from 'react';
-import { Alert } from 'src/components/patternfly-wrappers/l10n';
+import { Alert } from 'src/components';
 
 interface IProps {
   cancelAction: () => void;
